requirements.txt
streamlit
langchain
langchain_openai
langchain_community
faiss-cpu
pypdf
python-dotenv